requirements.txt
imgkit
ax-platform
pandas
slack-progress
slackclient
torchvision
opencv-python
tqdm
scipy
numpy
torch
wget
